move Welcome -> HomeAdministrator and HomeAdministrator -> ViewAssignments
